Fixed Snipping Tool Bug
Previously, the AHK version we were running had an internal issue
with the Mouse Hook. This problem caused click handlers to block
any mouse events after the mouse button was used in a
Snipping Tool (and possibly other types of) operation.

Now, with the newest compiler, the issue appears to have been
resolved (snipping tool is functional)